Fix an off-by-one error in the computation of days since 0th January 2000.
Why did this error go unnoticed? Because the test file up to version 0.0501
only tested the 21th June and at this time, the sunrise and the sunset
do not vary by much. So, computing the sunrise for 20th June and
comparing with the real value for 21st June would give very similar
results. But with dates far from the solstices, the sunrise and sunset
variation is much greater. So, even with a fudge value of 3 minutes,
the errors were noticeable.