Preventing useless computations of sunrise / sunset.
In some cases, the _sunset sub computes both sunrise and sunset, yet
the caller sub discards one of these values. With the new version,
the to-be discarded value is not computed when using the precise algorithm.

No such optimisation for the basic algorithm, we would not gain much, both
in terms of  CPU time and in terms of printed trace lines.